requirement.txt
django
djangorestframework
django
revauth
revcore
datatime